Reorganized continual_animation folder
