+++ mobject/image_mobject.py
@@ -7,8 +7,9 @@ import os
 from PIL import Image
 from random import random
 
-from mobject.mobject import Mobject
 from constants import *
+
+from mobject.mobject import Mobject
 from mobject.point_cloud_mobject import PMobject
 from utils.bezier import interpolate
 from utils.color import color_to_int_rgb
