+++ old_projects/mug.py
@@ -288,7 +288,7 @@ class AboutToyPuzzles(UtilitiesPuzzleScene, TeacherStudentsScene, ThreeDScene):
             eulers.get_bottom(),
             color = WHITE
         )
-        self.add(AmbientRotation(cube, axis = UP))
+        self.add(ContinualRotation(cube, axis = UP))
         self.play(
             GrowArrow(arrow_to_eulers),
             Write(eulers),
@@ -1884,7 +1884,7 @@ class EulersFormulaForGeneralPlanarGraph(LightUpNodes, ThreeDScene):
         self.play(FadeOut(self.vertices))
         self.play(ReplacementTransform(regions, cube, run_time = 2))
         cube.sort_submobjects(lambda p : -p[2])
-        self.add(AmbientRotation(cube, axis = UP, in_place = False))
+        self.add(ContinualRotation(cube, axis = UP, in_place = False))
         self.wait(3)
         self.play(
             FadeOut(self.top_formula),
